requirements.txt
langchain==0.2.10
langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-community==0.2.10
bs4==0.0.2
unstructured==0.15.0
pypdf==4.3.1
wikipedia==1.4.0
huggingface-hub==0.34.4 
sentence-transformers==5.1.0 
tokenizers==0.21.4 
transformers==4.55.4
langchain-chroma==0.1.2
langchain-core
flask
jinja2
faiss-cpu
python-dotenv
-e .